style(kitty): made default explicit
The setting in question:

```Nix
config.programs.kitty.shellIntegration.mode = "no-rc";
```

needs to have this as the default, or kitty would try to modify `~/.bashrc`.
That is why I felt it should be explicit, and not commented out as if it has
no effect when this is the required default.