feat: enabled auto cd in bash
this will allow you to `CD` into a directory when this is the argument you
provide to the shell. So running `./dir` will expand into `cd ./dir`. Same
with `/etc/dir` will expand into `cd /etc/dir`. This is a nice feature if
you forgot to prefix `cd` a lot.